// src/api/rathena/DatabaseDriver.types.ts
export const MvplogEntityType = zod.object({
  kill_char_id: zod.number(),
  map: zod.string(),
  monster_id: zod.number(),
  mvp_date: zod.date(),
  mvp_id: zod.number(),
  mvpexp: zod.string(),
  prize: zod.number(),
});